requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
faster-whisper==1.2.1
google-generativeai==0.8.5
python-multipart==0.0.20
python-dotenv==1.2.1